validation: Require there be non-bootstrap chunks in systems
Bootstrap chunks don't make it into the final system, so there needs to
be an extra check for empty systems after the sources have been collected.

This was complicated slightly by the fact that if you try to build a chunk
directly you will have no strata in your sources, hence no non-bootstrap
chunks, but validation for having been told to build a chunk is best
handled later.

This amends the old yarns that depended on building a bootstrap chunk
and adds a new one that explicitly builds a system with bootstrap
chunks.