unit tests: Fix invalid morphologies
Later validation work causes the morphologies to be validated, when they
weren't previously. This would cause the test suite to not pass, since
the morphologies defined in the tests are malformed.

One common problem was tests that, instead of a name field, had the name
of the morpholgy in a field called "chunk".

There were a few cases of new fields being needed, since the tests were
written before they became mandatory.

The most interesting failure was a Source being created, which instead
of being passed a morphology object, was passed a string.