Split chunk morphologies according to new rules
Filenames are now matched before chunks are constructed, so
bins.create_chunk now takes a list of relative file names.

bins.chunk_contents is gone, since this is now handled by passing
source.split_rules.partition the file names.

We now don't consider it to be a problem for directories to remain in
the DESTDIR after artifacts have been removed, since we need to handle
file matches implying their parent directories, and explicit matches
against directories.

NOTE: The bins_tests were broken in this patch, and are fixed in the
next. This was done to try and aid readability of the patch series.

Full functionality is still broken until stratum splitting is fixed.