Add optional overwrite option
This option lets the install-files config extension overwrite existing files.
A file will only be overwritten if the overwrite flag is specified for that file.

Since the overwrite arg is optionally prepended to the manifest line,
this patch should not break existing manifests

With this patch default config files can be replaced with project specific
config files