Use given artifact when constructing Strata
For some reason we used to create a new Artifact object with the same
name as the Stratum morphology for our cache key.

This is non-sensical, since we already have an Artifact object and it
breaks splitting strata.

NOTE: cmdtest tests do not pass, since they list files and artifacts
produced, which has changed since the new default splitting rules were
added. The next patch fixes this, but was kept as a separate commit for
readability.