yarns: Add test for not building unnecessary chunks
It's an unintended, but possibly not unwelcome feature, that if you have
a system that includes stratum artifacts that don't require building the
whole stratum, that those chunks are not built.

It's debatable whether the confusion of being able to have chunks
defined that aren't built outweighs reduced build-times, indeed the test
works by having a chunk that can never successfully build, which shows
that one pitfall is the possibility of not noticing some chunks fail to
build.

However, whether it's a good thing or not, we should document this
behaviour with a test, so that we don't accidentally break behaviour
that someone depends on.