cache-key: Fix obscure cache key collision
If two systems with the same name (e.g. different repo/ref) depend on the
same strata, then it will collide with systems which depend on different
artifacts from that stratum, but the same number of artifacts.

For example, if you checkout an existing branch and change the artifacts
used by one of its strata, then your local changes won't be built.

This is because the 'kids' field lists artifacts it depends on by their
cache-key, which is now no longer sufficient to uniquely identify
artifacts. The same number of artifacts issue is from it listing cache
keys multiple times.

The fix for this is to include the artifact name, so the 'kids' field is
now a list of dicts, with artifact name and cache key.

This is a dict rather than a tuple so that the generated /baserock
metadata is more readable.