Fix push_build_branches
A semantic error in the BuildBranch class meant that it was not possible
to push temporary branches.

This escaped testing since BuildBranch interacts too tightly with
other components to be easily unit-tested, so testing was deferred to
a yarn test.

However, coverage isn't measured in yarn tests, so this code path was
forgotten.