Pass in user's PYTHONPATH to morph when run from yarns
We carefully _add_ to PYTHONPATH in ./check, if it was set by the user.
However, yarn cleans the environment when it runs tests, so we tell it
to add PYTHONPATH from ./check to the test environment.

Additionally, we change yarns/morph.shell-lib so it doesn't override
PYTHONPATH, but adds to it.

All of this is necessary to get morph, when run by yarn steps, to have
the right PYTHONPATH, which can be (and currently is) to allow the user
to specify un-installed versions of dependencies, such as cliapp.