Reduce spam from setting up staging area
Make the message "Installing chunk..." be chatty, i.e., only displayed
when the user turns verbosity higher. Most of the time the chunk is
already unpacked in the cache, so installing it takes a fraction of
a second.

Add a new message, at default verbosity, when a chunk needs to be
unpacked.  This can take a while, so a message is appropriate, so the
user knows what is happening.