Give a useful error when attempting to build a cluster morph
Previously this resulted in a confusing traceback.

This is a quick fix. I'd rather insert the error in the buildcommand
module instead, but that code assumes the source being built is a
system in several places before it actually checks the kind. Those
would all need to be changed, or the code would need to reworked to
call _validate_root_kind() much earlier.

The Application.traverse_morphs() method is rather ugly anyway, so I'm
happy to add further ugliness to it for the time being.