WorkerConection: _handle_exec_response
The exec_response_msg also needs to be sent to a number of initiators,
so we give it a list of ids not just one.

The exec_response_msg will be sent to the controller once the artifacts
have been cached successfully.

There's no longer any need to use a route map to retrieve
the id of the initiator, since this is stored with the job