Use logfile kwarg to generate build log
We could just set stdout to subprocess.PIPE
then read from the pipe, but then we won't get the output
till the command's finished and some commands take a long time.

Using the logfile kwarg a file will be created by tee
and the output will be written to it in 'real time'.