Make Jobs finish when caching is complete
If a new build request makes a request for an artifact that is currently being
cached then the artifact will be needlessly rebuilt.
To avoid this the new build request should wait for caching to finish.

We rename _ExecStarted, _ExecEnded, _ExecFailed to
_JobStarted, _JobFinished, _JobFailed
and Job's is_building attribute is renamed to running.