Make uuid mandatory when calling create_btrfs_system_layout
It's only ever called from the core of the write extensions library, so
we can change it to be mandatory in all cases.

install_extlinux is left with the uuid being optional and defaulting to
/dev/sda, since it is called directly from the rawdisk write extension
currently, and upgrades of systems that have an initramfs will be part
of a later patch series.