Simplify morph edit syntax and logic
Since we removed ref: fields from the system morph files, some of the logic
in morph edit is no longer needed. In particular, running

  morph edit <system> <stratum>

is a no-op. This is because the version of <system> and <stratum> are now
implicit from the context of what we are doing. In other words we're always
working with the current version of <system> and <stratum> from the system
branch we are in.

Because of the complexity of morph's logic, we didn't notice this when
dropping the ref: fields, and we missed the opportunity to simplify our
logic for 'morph edit'

This patch aims to provide the simplest possible implementation of
morph edit:

 morph edit <chunk>

It checks all strata for instances of <chunk>, and does what morph edit should
do for the instances it finds.

A later patch can add warnings to help users deal with situations where <chunk>
is not found, or is found more than once.

Also since this changes the syntax of morph, it breaks many of our tests. Later
patches will address that.