Fix: force select to wait on timeout
Currently the mainloop unintentionally busy-waits if
we can't connect to the controller.

We want the mainloop's select call to wait for the timeout
not for this socket's descriptors (which are always ready).

We could just call stop_reading() and stop_writing() but since
we won't be needing this socket again we may as well close the
entire event source, which calls stop_reading(), stop_writing()
and then closes the socket.