Use abbreviated commit SHA as version number, rather than the tag
This still (ab)uses 'git describe', in order to get an '-unreproducible'
postfix when the tree has uncommitted changes.

Previously we tagged morph.git with the name of the corresponding
Baserock release that it was to be part of. This made the Baserock
release process awkward. We needed to first create and test the
release images, then (once we knew the version of Morph we were
releasing was good) tag morph.git, and then build all the release
images again so that the new tag would be picked up and the output of
`morph --version` would be correct.