Only updated cached git repos when necessary
When the given ref points to a specific commit, and it's already available in the
locally cached copy of repo, there's no need to update the repo.

If the ref points to a branch or tag, and the user didn't pass --no-git-update, the
locally cached copy of the repo will still be updated.

This should speed up many Morph commands.