Remove some cache key related log messages
There's no need to log every time we look something up in a dict. This just
makes log files huge.

The CacheKeyComputer.compute_key() function still logs the first time it
calculates a cache key for an Artifact instance. This message now includes
the hex string that is used to identify the artifact.