Make our use of json binary path safe
json only accepts unicode. Various APIs such as file paths and environment
variables allow binary data, so we need to support this properly.

This patch changes every[1] use of json.load or json.dump to escape
non-unicode data strings. This appears exactly as it used to if the
input was valid unicode, if it isn't it will insert \xabcd escapes in
the place of non-unicode data.

When loading back in, if json.load is told to unescape it with
`encoding='unicode-escape'` then it will convert it back correctly.

This change was primarily to support file paths that weren't valid
unicode, where this would choke and die. Now it works, but any tools
that parsed the metadata need to unescape the paths.

[1]: The interface to the remote repo cache uses json data, but I haven't
     changes its json.load calls to unescape the data, since the repo
     caches haven't been made to escape the data.