Remove update-gits command
This hasn't been used in a long time to my knowledge, its API completely
misses the point, and its implementation relies on old APIs that need to
change.

Until we discover we need it, and work out what it should look like, I
think the best thing to do is get rid of it.