Rewrite traverse_morphs to be able to load chunk morphologies from definitions
It will now load the morphology from the definitions repository if the
"morph" field is present in the chunk spec.

Rather than adapting the loop to fit yet-more changing circumstances, it
has been partially rewritten, so there is one loop for loading
systems and strata from definitions.git, another for chunks from
definitions.git, and a third for chunks in the source repository.

This is tidier than attempting to fit the logic in the main loop, as it
makes it easier to remove afterwards when we no longer need to load
chunk morphologies from the source repository.