StagingArea: Ensure staging are paths are bytestrings
If the destdir path returned when creating a staging area is a unicode
string, then when attempting to `os.walk(destdir)`, it will encounter
unicode errors if there are file paths in the destdir that are not
representable as unicode strings.

For various as-yet unknown reasons, when building stage-2 eglibc it
produces file paths that are not unicode compatible.

There was previously a patch to fix this issue with regards to creating
the metadata files, but it did not fix all the issues, because the build
at the time was local rather than distributed.

This is failing during a distributed build because morphologies are
serialised into json, and during deserialisation their string values are
left as unicode.

Rather than doing the byte-string conversion during deserialisation, I
have chosen to do it when the contents of the morphology are used,
because it's only at the point where it's used to create a file path,
that it matters whether it's unicode or not.