Fix not including symlinks in chunks if they exist on the host
With os.walk, if the target of the link doesn't exist, or it is a link
to a file, it ends up in the basenames list.

If it is a link to a directory, it goes in the subdirs list.

There's a bug in the subdirsymlinks check, in that it checks if the
wrong file is a symlink, so it never returns them.

This was missed, since we did not have the cross bootstrap in CI.

This is not eligible for our yarn tests, since to trigger this would
require changes to the host system's rootfs, so it's a system-level test.

To test this properly requires putting the cross bootstrap in CI.