Cut BuildBranch out of morphlib.extensions
This was previously used just so it could get the right repo and ref to
read the file out of.

However, there was a subtle bug in this behaviour, as if we had not
previously used morph build in that branch, it would attempt to read the
extensions from a branch which didn't exist.

So now it reads it from the working tree, which always exists.