Treat untracked morphologies as uncommitted changes
We used to coincidentally include morphologies because we later include
every morphology in the build graph into our temporary build branches.

Since we're now checking whether there's any uncommitted changes before
attempting to create a temporary build branch, this means that we can no
longer build uncommitted morphologies if they aren't reported as
changes.

So this patch makes an exception to the untracked changes rule for
anything that ends with .morph.

It's still confusing that some files aren't included in temporary build
branches, but that would cause performance regressions, so we'll limit
it to just morphologies for now, until we make a decision on what
uncomitted content we care about.