Make morph show-branch-root print the path
The help for the show-branch-root command said it returns a path, but
the command and the yarns just showed the aliased url it was cloned
from.

Given I found myself needing the path in some scripts, not the repo url,
I think it's more useful to reconcile the difference this way.