Fix `morph edit` when repo has the same ref as system branch
There was a check in it to see whether it needed to do the git branch
and git checkout based on whether the name of the branch matched that in
the morphology.

This had a couple of problems:

1.  Now that we aren't always building from HEAD, we need to be able to
    roll its commit back, so using the existing branch isn't always the
    best idea.
2.  It only checks the "ref" field, not "unpetrify-ref", so even though
    we clone the right ref in there, it's checking the commit id against
    the system branch name, so would always try to re-create the branch,
    and fail when it already exists.

So now, we remove the original ref and re-create it with our preferred
HEAD.

A better solution might be to change the clone logic to not
automatically checkout HEAD, and instead require an explicit branch then
checkout, but the initial clone logic is shared with build, and I didn't
feel like tracking down all the different places that it was used.