Only autodetect morphology when result is 404
The MorphologyFactory class will use a RemoteRepoCache to see if a
morphology file exists, and if it doesn't, uses a file listing to see
if it can detect what build-system is uses, hence what the default
morphology should be.

However, it was overly generic in what error cases it would accept as
the morphology not being found, so if the RemoteRepoCache was suddenly
un-resolvable for a brief period, then it would assume the morphology
didn't exist, and use the default one.

This happened to a user, and the result was a full rebuild.

So we now fix this by only raising the exception that means the file
didn't exist, if we got a HTTP 404.