deploy: Check correct usage of --upgrade for rawdisk deployments
This avoids confusion when the user expected to be doing an initial
deployment and wasn't aware that a file with the same name as the
target already existed. Previously rawdisk.write would try to mount
the file and upgrade it.

Now we require the user to pass '--upgrade' when they intend to
upgrade, as with other deployment extensions.