Prevent git-replace refs affecting git operations
We assumed that the sha1 of the tree of the commit of the ref we care
about was sufficient to cache, but `git replace` means that you need to
know the state of other branches too, since anything in refs/replace can
completely change what the tree you check-out is.

This behaviour can be disabled globally by setting
GIT_NO_REPLACE_OBJECTS, so we're going to do that.

If we need to integrate a project that uses git-replace to change the
contents of their git trees then we could support that by:

    if any(refs/replace/*):
        potentially_replacable_objects = [
            `git rev-parse HEAD`,
            `git rev-parse HEAD^{commit}`,
            `git rev-parse HEAD^{tree}`]
        potentially_replacable_objects.extend(
            `git ls-tree -r HEAD | awk '{print $3}'`)
        # NOTE: doesn't handle submodules, and you'd need to expand this
        # set whenever you process a replacement
    for object in refs/replace/*:
        if basename(object) not in potentially_replacable_objects: continue
        cache_key['replacements'][basename(object)] = `git rev-parse $object`

If we were to support this would need to traverse the tree anyway, doing
replacements, so we may as well use libgit to do the checkout anyway,
and list which replacements were used.

However, since the expected use-case of `git replace` is as a better way
to do history grafting, we're unlikely to need it, as it would only have
any effect if it replaced the commit we were using with a different one.

Rubber-stamped-by: Daniel Silverstone