yarns: Use a more realistic simulation of a git server
Previously we would use file: URIs to point to the git repositories.

This was fast and simple, but had the drawback that it bypassed all the
git cache logic, so changes to the git cache weren't adequately covered
by the test suite.

Now we spool up a simulated git server per scenario, and shut it down at
the end.