Use git's remotes to determine unpushed branches
Rather than using `git ls-remote` every time to see if there are changes
at the remote end, use a local cache.

Git already solves this problem with its refs/remotes/$foo branch
namespace, so we can just use that instead.

In addition, the detection of which upstream branch to use has been
improved; it now uses git's notion of which the upstream branch of your
current branch is, which saves effort in the implementation, and allows
the name of the local branch to differ from that of the remote branch.

This now won't notice if the branch you currently have checked out had
commits pushed from another source, but for some use-cases this is
preferable, as the result equivalent to if you had built before the
other push.

It may make sense to further extend this logic to check that the local
branch is not ahead of the remote branch, instead of requiring them to
be equal.