Remove cmdtest for builds producing a separate -kernel artifact
The notion of creating a separate kernel artifact comes from before we
had deployments.

It was so that we could boot a system in qemu without having to extract
the kernel from the rootfs. Now we can do this by copying the kernel out
at deployment time, so we don't really need to do this any more.