Eliminate BuilderBase.new_artifact
It was an odd thing to have, when Artifact objects are part of your
input.

Its purpose appears to have been to allow the build step to produce an
artifact called $morphology_name-rootfs, but since the split rules
decide that the artifact is called that anyway, the new_artifact step is
redundant.