Move metadata_version value to cachekeycomputer
metadata_version is for when the format of the /baserock metadata files
changes.

This means it would make sense for it to either live with the code for
generating the metadata, or the cache key code so it lives with the rest
of the compatibility values.

Since the code for generating the metadata isn't in a nice module
anywhere, I've put it in the cachekeycomputer module.