Make morphologies hashable
I need to be able to deduplicate a list of morphologies. Putting it in a
set is the easiest way, but it needs to be hashable.

It's not included in dicts by default, since they're stored by
reference, and you can change them while they're in the dict, so the
hash value can change.

I don't need to deduplicate morphologies by their contents, just by
reference though, so using `id` as the hash function is sufficient.