Add test shell for use in staging area tests
This is intended to be statically linked, so it can be run in the
staging area, without having to build a libc.

You shouldn't generally statically link GLibc, because it dynamically
links things like NSS modules or locale data, but we only need some very
simple stuff, so we can get away with that.

There's also potential licensing issues, as GLibc is LGPLv2, so
distribution requires providing the ability to re-link against another
library, but its use in the test-suite shouldn't count as distribution.

There's a couple of commands, the only two needed by the yarn test
suite are "copy files" which is like `cp -r . "$DESTDIR"`, and "false",
which is for commands to deliberately fail.