distbuild: allow daemons to bind to ephemeral ports
You can bind to an ephemeral port by passing 0 as the port number.
To work out which port you actually got, you need to call getsockname().

To facilitate being able to spawn multiple copies of the daemons for
testing environments, you can pass a -file option, which will make the
daemon write which port it actually bound to.

If this path is a fifo, reading from it in the spawner process will
allow synchronisation of only spawning services that require that port to
be ready after it is.