deploy extensions: Don't crash if someone builds at the same time
If a build happens, it creates a new network namespace, and if this
happens while you have a disk image mounted, then you can't remove the
mount-point, because the other namespace is using it.

We can avoid the other namespace keeping this mount-point open by
creating the disk image in a private mount namespace, so it never sees
it.

The nicest way to do this is to have every extension run in a private
mount namespace, since you'd have to have extensions re-exec themselves,
since the appropriate system calls aren't exposed very well.