build: Resolve the ref being built in the user's local definitions.git
Most usefully, this patch means that Morph no longer updates its cached
copy of definitions.git every time you run `morph build`.

Also, it prevents confusion in the following situation. Imagine I have
run:

    morph checkout baserock:baserock/definitions master

I then wait a while, during which time someone pushes to 'master' in the
definitions.git repo that I cloned from. Now I run:

    cd master
    morph build systems/whatever.morph

Which commit does it build, the local head of 'master' or the remote
head of 'master'?

The answer, both before and after this patch, is that it builds the
local version of master. But previously, this only happened because of
the magic that we have to detect local changes. With this patch, the
local change detection could be disabled and `morph build` would still
build what the user had checked out as 'master' locally, not whatever
'master' pointed to in the remote repo.